+++ src/felupe/mechanics/_solidbody_incompressible.py
@@ -23,9 +23,8 @@ import numpy as np
 
 from ..assembly import IntegralForm
 from ..constitution import AreaChange
-from ..field import Field, FieldAxisymmetric, FieldContainer
-from ..math import ddot, det, dot, dya, rotate_points, transpose
-from ..region import RegionHexahedron, RegionQuad
+from ..field import FieldAxisymmetric
+from ..math import ddot, det, dot, dya, transpose
 from ._helpers import Assemble, Evaluate, Results, StateNearlyIncompressible
 from ._solidbody import Solid
 
@@ -431,7 +430,7 @@ class SolidBodyNearlyIncompressible(Solid):
             "results.statevars": self.results.statevars.copy(),
         }
 
-    def restore(self, checkpoint, restore_statevars=True):
+    def restore(self, checkpoint, restore_statevars=True, restore_state=True):
         """Restore a checkpoint inplace.
 
         Parameters
@@ -449,9 +448,11 @@ class SolidBodyNearlyIncompressible(Solid):
         """
 
         self.field.restore(checkpoint)
-        self.results.state.u[:] = checkpoint["results.state.u"]
-        self.results.state.p[:] = checkpoint["results.state.p"]
-        self.results.state.J[:] = checkpoint["results.state.J"]
+
+        if restore_state:
+            self.results.state.u[:] = checkpoint["results.state.u"]
+            self.results.state.p[:] = checkpoint["results.state.p"]
+            self.results.state.J[:] = checkpoint["results.state.J"]
 
         if restore_statevars:
             self.results.statevars[:] = checkpoint["results.statevars"]
@@ -485,34 +486,7 @@ class SolidBodyNearlyIncompressible(Solid):
 
         """
 
-        # revolve the mesh around the x-axis and create new region and new field
-        new_mesh = self.field.region.mesh.revolve(n=n, phi=phi, axis=0, expand_dim=True)
-        new_region = {
-            RegionQuad: RegionHexahedron,
-        }[
-            type(self.field.region)
-        ](new_mesh)
-
-        if np.isscalar(phi):
-            rotation_angles = np.linspace(0, phi, n)
-        else:
-            rotation_angles = phi
-            n = len(rotation_angles)
-
-        new_values = []
-        for angle_deg in rotation_angles:
-            new_values.append(
-                rotate_points(
-                    points=np.pad(self.field[0].values, ((0, 0), (0, 1))),
-                    angle_deg=angle_deg,
-                    axis=0,
-                )
-            )
-
-        dim = new_values[-1].shape[1]
-        new_field = FieldContainer(
-            [Field(new_region, dim=3, values=np.array(new_values).reshape(-1, dim))]
-        )
+        new_field = self.field.revolve(n=n, phi=phi)
 
         # create a new solid body
         new_solid = SolidBodyNearlyIncompressible(